backend/requirements.txt
# Core Django
Django==4.2.10
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
drf-spectacular==0.27.1

# CORS
django-cors-headers==4.3.1

# Environment management
python-dotenv==1.0.1

# File handling
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
